classes abstractas creadas
