+++ src/Library/Characters/ICharacter.cs
@@ -2,24 +2,75 @@ using System.Collections.Generic;
 namespace RoleplayGame
 {
     public abstract class ICharacter
-    {
-        private int health = 100;
+    {   /*Le agregue la etiqueta publica todas las caracteristicas de esta clase abstracta para que las otras clases pudieran tener acceso*/
+        public int health = 100;
 
-        private List<IItem> items = new List<IItem>();
-        string Name { get; set; }
+        public List<IItem> items = new List<IItem>();
+        public string Name { get; set; }  
+        public int AttackValue
+        {
+            get
+            {
+                int value = 0;
+                foreach (IItem item in this.items)
+                {
+                    if (item is IAttackItem)
+                    {
+                        value += (item as IAttackItem).AttackValue;
+                    }
+                }
+                return value;
+            }
+        }
 
-        int Health { get; }
+        public int DefenseValue
+        {
+            get
+            {
+                int value = 0;
+                foreach (IItem item in this.items)
+                {
+                    if (item is IDefenseItem)
+                    {
+                        value += (item as IDefenseItem).DefenseValue;
+                    }
+                }
+                return value;
+            }
+        }
+        public int Health
+        {
+            get
+            {
+                return this.health;
+            }
+            private set
+            {
+                this.health = value < 0 ? 0 : value;
+            }
+        }
 
-        int AttackValue { get; }
+        public void AddItem(IItem item)
+        {
+            this.items.Add(item);
+        }
 
-        int DefenseValue { get; }
+        public void RemoveItem(IItem item)
+        {
+            this.items.Remove(item);
+        }
 
-        void AddItem(IItem item);
+        public void Cure()
+        {
+        this.Health = 100;
+        }
 
-        void RemoveItem(IItem item);
-
-        void Cure();
-
-        void ReceiveAttack(int power);
+        public void ReceiveAttack(int power)
+        {
+            if (this.DefenseValue < power)
+            {
+                this.Health -= power - this.DefenseValue;
+            }
+        }
     }
 }
